refactor(tau-gateway): split safety runtime module (#2679)
